Refactorings for Microservices
