--- services/APIGateway/src/main/java/com/example/APIGateway/Authentication/Core/DataTypeObjects/AuthenticationRequest.java
@@ -1,16 +1,0 @@
-package com.example.APIGateway.Authentication.Core.DataTypeObjects;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationRequest {
-    private String email;
-    String password;
-}
