--- services/APIGateway/src/main/java/com/example/APIGateway/Authentication/Core/DataTypeObjects/ErrorResponse.java
@@ -1,14 +1,0 @@
-package com.example.APIGateway.Authentication.Core.DataTypeObjects;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class ErrorResponse {
-    private String error;
-}
