--- services/APIGateway/src/main/java/com/example/APIGateway/Authentication/Core/DataTypeObjects/TokenRequest.java
@@ -1,15 +1,0 @@
-package com.example.APIGateway.Authentication.Core.DataTypeObjects;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class TokenRequest {
-    private String refresh;
-}
