--- services/APIGateway/src/main/java/com/example/APIGateway/Authentication/Core/Enums/Role.java
@@ -1,52 +1,0 @@
-package com.example.APIGateway.Authentication.Core.Enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-import static com.example.APIGateway.Authentication.Core.Enums.Permission.*;
-
-@RequiredArgsConstructor
-public enum Role {
-
-    USER(Collections.emptySet()),
-    ADMIN(
-            Set.of(
-                    ADMIN_READ,
-                    ADMIN_UPDATE,
-                    ADMIN_DELETE,
-                    ADMIN_CREATE,
-                    DEVELOPER_READ,
-                    DEVELOPER_CREATE,
-                    DEVELOPER_UPDATE,
-                    DEVELOPER_DELETE
-            )
-    ),
-    MANAGER(
-            Set.of(
-                    DEVELOPER_READ,
-                    DEVELOPER_CREATE,
-                    DEVELOPER_UPDATE,
-                    DEVELOPER_DELETE
-            )
-    )
-
-    ;
-
-    @Getter
-    private final Set<Permission> permissions;
-
-    public List<SimpleGrantedAuthority> getAuthorities() {
-        var authorities = getPermissions()
-                .stream()
-                .map(permission -> new SimpleGrantedAuthority(permission.getPermission()))
-                .collect(Collectors.toList());
-        authorities.add(new SimpleGrantedAuthority("ROLE_" + this.name()));
-        return authorities;
-    }
-}
