--- services/APIGateway/src/main/java/com/example/APIGateway/Authentication/Entity/UserEntity.java
@@ -1,70 +1,0 @@
-package com.example.APIGateway.Authentication.Entity;
-
-import com.example.APIGateway.Authentication.Core.Enums.Role;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-import org.springframework.data.mongodb.core.mapping.Field;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-import java.util.List;
-
-@Document(collection = "users")
-@Builder
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserEntity implements UserDetails {
-    @Id
-    private String id;
-    @Field
-    private String username;
-    @Field
-    private String password;
-    @Field
-    private String email;
-
-    @Field
-    private Role role;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of(new SimpleGrantedAuthority(role.name()));
-    }
-
-    @Override
-    public String getPassword() {
-        return password;
-    }
-
-    @Override
-    public String getUsername() {
-        return email;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
