--- services/APIGateway/src/main/java/com/example/APIGateway/Authentication/Repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.example.APIGateway.Authentication.Repository;
-
-import com.example.APIGateway.Authentication.Entity.UserEntity;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import java.util.Optional;
-
-public interface UserRepository  extends MongoRepository<UserEntity,String> {
-    Optional<UserEntity> findByEmail(String email);
-}
