+++ services/APIGateway/src/main/java/com/example/APIGateway/Configuration/ApplicationConfiguration.java
@@ -1,7 +1,7 @@
 package com.example.APIGateway.Configuration;
 
 
-import com.example.APIGateway.Authentication.Repository.UserRepository;
+import com.example.APIGateway.Repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
