+++ services/APIGateway/src/main/java/com/example/APIGateway/Core/DataTypeObjects/ErrorResponse.java
@@ -1,0 +1,14 @@
+package com.example.APIGateway.Core.DataTypeObjects;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class ErrorResponse {
+    private String error;
+}
