+++ services/APIGateway/src/main/java/com/example/APIGateway/Core/DataTypeObjects/RegisterRequest.java
@@ -1,0 +1,16 @@
+package com.example.APIGateway.Core.DataTypeObjects;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class RegisterRequest {
+    private String username;
+    private String password;
+    private String email;
+}
