+++ services/APIGateway/src/main/java/com/example/APIGateway/Repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.example.APIGateway.Repository;
+
+import com.example.APIGateway.Entity.UserEntity;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import java.util.Optional;
+
+public interface UserRepository  extends MongoRepository<UserEntity,String> {
+    Optional<UserEntity> findByEmail(String email);
+}
