+++ services/Goals/src/main/java/com/example/server/Core/DataTransferObjects/ErrorResponse.java
@@ -1,0 +1,14 @@
+package com.example.server.Core.DataTransferObjects;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class ErrorResponse {
+    private String error;
+}
