--- services/Goals/src/main/java/com/example/server/Goals/Core/DataTransferObjects/ErrorResponse.java
@@ -1,14 +1,0 @@
-package com.example.server.Goals.Core.DataTransferObjects;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class ErrorResponse {
-    private String error;
-}
