--- services/Goals/src/main/java/com/example/server/Goals/Core/DataTransferObjects/TaskRequest.java
@@ -1,20 +1,0 @@
-package com.example.server.Goals.Core.DataTransferObjects;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import java.util.Date;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class TaskRequest {
-
-    private String Title;
-    private String Description;
-    private String Status;
-    private Date StartTime;
-    private Date EndTime;
-    private String Notes;
-}
